Рефакторит дирректорию типов
