+++ project/src/mocks/reviews.ts
@@ -1,4 +1,4 @@
-import {Review} from '../types/reviews';
+import {Review} from '../types/data';
 import {USERS} from './users';
 
 export const REVIEWS: Review[] = [
