+++ project/src/pages/room/room.tsx
@@ -6,8 +6,8 @@ import NotFound from '../not-found/not-found';
 import HostUser from '../../components/host-user/host-user';
 import PropertyReviews from '../../components/property-reviews/property-reviews';
 import PremiumLabel from '../../components/premium-label/premium-label';
-import {Offer} from '../../types/offers';
-import {Review} from '../../types/reviews';
+import {Offer} from '../../types/data';
+import {Review} from '../../types/data';
 import {calculateRatingStars} from '../../utils';
 import styles from './room.module.scss';
 import Map from '../../components/map/map';
