--- project/src/types/offers.ts
@@ -1,31 +1,0 @@
-import {User} from './users';
-
-export type Location = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-}
-
-export type City = {
-  location: Location;
-  name: string;
-}
-
-export type Offer = {
-  id: number;
-  bedrooms: number;
-  city: City;
-  description: string;
-  goods: string[];
-  host: User;
-  images: string[];
-  isFavorite: boolean;
-  isPremium: boolean;
-  location: Location;
-  maxAdults: number;
-  previewImage: string;
-  price: number;
-  rating: number;
-  title: string;
-  type: string;
-}
