--- project/src/types/reviews.ts
@@ -1,9 +1,0 @@
-import {User} from './users';
-
-export type Review = {
-  id: number;
-  comment: string;
-  date: Date;
-  rating: number;
-  user: User;
-}
