--- project/src/types/users.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id: number;
-  name: string;
-  avatarUrl: string;
-  isPro: boolean;
-}
